feat:【PAC模板】流水线模板支持PAC特性 #11414 feat:【PAC模板】新建流水线模板 #10336

+++ src/backend/ci/core/common/common-pipeline/src/main/kotlin/com/tencent/devops/common/pipeline/ITemplateFunction.kt
@@ -37,7 +37,7 @@ interface ITemplateFunction {
     @get:Schema(title = "模板ID", required = false)
     var templateId: String?
     @get:Schema(title = "版本", required = false)
-    var templateVersion: Int?
+    var templateVersion: Long?
     @get:Schema(title = "模板参数构建", required = false)
     var templateParams: List<BuildFormProperty>?
 }
